Use a hardcoded data directory for tests.
This is less clean than using a fresh tempdir each time, and less
portable because we're hardcoding a Unix-style path, but it means we
won't keep re-downloading the same files from travellermap every time we
run "cargo test", which I think is more important than the other
factors.

If we need a clean test, we can just delete the test directory
and everything will re-download on the nest test run.