Find endpoint trade by converting BTN -> credits -> BTN
This requires breaking out of the nested loops in populate_trade_routes
less aggressively, because a pair of worlds that doesn't quite reach
BTN 8 for a minor trade route might still generate enough credits of
trade to matter.  So lower the threshold for breaking out to 7 and
add a TRIVIAL_ROUTE_THRESHOLD constant to be consistent.