Parallelize finding route paths
World.find_route_paths was accumulating its results in a couple of
mutably-referenced maps (pairs of world coords to credits and individual
worlds to transient credits) for convenience.  Now it simply returns its
results in two local temporary maps, which get pushed onto a vector, and
we then reduce the vector, accumulating the results into the two final
maps.

On my 16-core 32-thread CPU, this cuts the time to find route paths by
about 80% and reduces the total runtime of the program by over 30%,
for a test with 30 sectors and 12056 worlds.