Make Sector.name the base name of the filenames
It turns out that some sectors don't have the original sector filename
as the first name, which meant I needed to track the original filename
separately to make sure I could find all the data files.  Now name is
the original filename, and names is the full list of names in arbitrary
order.  The name() method was removed since we can just use the name
field.