Convert distances to u16.
This saves enough memory to finish all.txt (142 sectors and 51264
worlds) on a 64 GB machine.  (It takes 2 minutes and 54 seconds of
wall clock time on my AMD 5950X CPU.)

u16 rather than i16 was required because i16::MAX is only 32767
and we want to support at least 51264 worlds.  This means that
the sentinal value NO_PRED_NODE in the predecessor table now has
to be in-band; it's set to u16::MAX - 1.

If in the future we want to support more than 2^16 worlds, we'll
need to go back to 32-bit numbers.  But then the program will use
more memory, and 51264 worlds will not quite fit in 64 GB.

Maybe the data type used should be configurable, but for now that
seems like unnecessary complexity.